requirements.txt
torch
torchvision
timm
albumentations
opencv-python
fastapi
uvicorn
streamlit
pillow
numpy
pandas
scikit-learn
python-multipart
requests